add SinCosEncoder class
